ptz: Move backend-specific code to each backend

+++ src/dummy-backend.hpp
@@ -15,5 +15,5 @@ public:
 	void set_pantilt_speed(int pan, int tilt) override;
 	void set_zoom_speed(int zoom) override;
 	void recall_preset(int preset) override;
-	int get_zoom() override;
+	float get_zoom() override;
 };
